+++ CHANGES.rst
@@ -4,5 +4,6 @@ CHANGES
 
 develop
 =======
+- INV-4068 Added authorization endpoints and save all available data about user to database
 - And so it begins